Add Revive method
Added revive method along with death and some other stuff. The revive method is a sequence that also plays an animation of flapping wings.
Added wings and halo to heirarchy and scene, used in revive.
Death is triggered on minHeight or 0 health and calls Revive after a set amount of time. still needs a way to show actual death animation then make character disappear for the time until revive.
Added IEnumerator to RTool that waits for secs then runs a UnityAction.
Changed enemyspawnerhandler to only spawn red and blue, should definitely use a FloatRange's min and max to set those in editor through reference instead of hardcode.