Preliminary Zoning
Put into place defining zones which dictate the movement of slimes.  Currently limited to a single range for horizontal and vertical (piecewise)

Usage: Main zone in the center, allows slimes to jump left and right.  Left and Right zones force slimes to move right and left, respectively in order to keep slimes within the zone (e.g. on the platform)

Example scene: TestB includes mostly examples of keeping slimes on their platforms, but also includes one example of a zone setup which will allow slimes to move to another platform.

TODO: Refine movement definition to allow for multiple movement vectors and/or minimum magnitudes
Maybe add a bias to the move if the player is nearby/in the zone